Moves & renames in search components
