--- plugin-hrm-form/src/components/search/CasePreview/index.js
@@ -1,3 +1,0 @@
-import CasePreview from './CasePreview';
-
-export default CasePreview;
