--- plugin-hrm-form/src/components/search/ContactPreview/index.js
@@ -1,3 +1,0 @@
-import ContactPreview from './ContactPreview';
-
-export default ContactPreview;
